tests: Move utility functions into own source file
